+++ .env
@@ -1,4 +1,4 @@
-OPERATOR_SDK_VERSION=v1.31.0
+OPERATOR_SDK_VERSION=v1.33.0
 REVIEWERS=vivekr-splunk,akondur
 GO_VERSION=1.23.0
 AWSCLI_URL=https://awscli.amazonaws.com/awscli-exe-linux-x86_64-2.8.6.zip
