+++ Makefile
@@ -135,10 +135,10 @@ test: manifests generate fmt vet envtest ## Run tests.
 ##@ Build
 
 build: setup/ginkgo manifests generate fmt vet ## Build manager binary.
-	go build -o bin/manager main.go
+	go build -o bin/manager cmd/main.go
 
 run: manifests generate fmt vet ## Run a controller from your host.
-	go run ./main.go
+	go run ./cmd/main.go
 
 docker-build: #test ## Build docker image with the manager.
 	docker build -t ${IMG} .
