+++ internal/controller/clustermanager_controller.go
@@ -14,15 +14,16 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package controllers
+package controller
 
 import (
 	"context"
 	enterpriseApi "github.com/splunk/splunk-operator/api/v4"
+	"github.com/splunk/splunk-operator/internal/controller/common"
 	"time"
 
 	"github.com/pkg/errors"
-	common "github.com/splunk/splunk-operator/controllers/common"
+	enterprise "github.com/splunk/splunk-operator/pkg/splunk/enterprise"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
@@ -35,19 +36,17 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/predicate"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 	"sigs.k8s.io/controller-runtime/pkg/source"
-
-	enterprise "github.com/splunk/splunk-operator/pkg/splunk/enterprise"
 )
 
-// SearchHeadClusterReconciler reconciles a SearchHeadCluster object
-type SearchHeadClusterReconciler struct {
+// ClusterManagerReconciler reconciles a ClusterManager object
+type ClusterManagerReconciler struct {
 	client.Client
 	Scheme *runtime.Scheme
 }
 
-//+kubebuilder:rbac:groups=enterprise.splunk.com,resources=searchheadclusters,verbs=get;list;watch;create;update;patch;delete
-//+kubebuilder:rbac:groups=enterprise.splunk.com,resources=searchheadclusters/status,verbs=get;update;patch
-//+kubebuilder:rbac:groups=enterprise.splunk.com,resources=searchheadclusters/finalizers,verbs=update
+//+kubebuilder:rbac:groups=enterprise.splunk.com,resources=clustermanagers,verbs=get;list;watch;create;update;patch;delete
+//+kubebuilder:rbac:groups=enterprise.splunk.com,resources=clustermanagers/status,verbs=get;update;patch
+//+kubebuilder:rbac:groups=enterprise.splunk.com,resources=clustermanagers/finalizers,verbs=update
 //+kubebuilder:rbac:groups=apiextensions.k8s.io,resources=customresourcedefinitions,verbs=get;list
 //+kubebuilder:rbac:groups=core,resources=services,verbs=get;list;watch;create;update;patch;delete
 //+kubebuilder:rbac:groups=core,resources=services/finalizers,verbs=get;list;watch;create;update;patch;delete
@@ -64,21 +63,22 @@ type SearchHeadClusterReconciler struct {
 // Reconcile is part of the main kubernetes reconciliation loop which aims to
 // move the current state of the cluster closer to the desired state.
 // TODO(user): Modify the Reconcile function to compare the state specified by
-// the SearchHeadCluster object against the actual cluster state, and then
+// the ClusterManager object against the actual cluster state, and then
 // perform operations to make the cluster state reflect the state specified by
 // the user.
 //
 // For more details, check Reconcile and its Result here:
 // - https://pkg.go.dev/sigs.k8s.io/controller-runtime@v0.10.0/pkg/reconcile
-func (r *SearchHeadClusterReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
-	reconcileCounters.With(getPrometheusLabels(req, "SearchHeadCluster")).Inc()
-	defer recordInstrumentionData(time.Now(), req, "controller", "SearchHeadCluster")
+func (r *ClusterManagerReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
+	// your logic here
+	reconcileCounters.With(getPrometheusLabels(req, "ClusterManager")).Inc()
+	defer recordInstrumentionData(time.Now(), req, "controller", "ClusterManager")
 
 	reqLogger := log.FromContext(ctx)
-	reqLogger = reqLogger.WithValues("searchheadcluster", req.NamespacedName)
+	reqLogger = reqLogger.WithValues("clustermanager", req.NamespacedName)
 
-	// Fetch the SearchHeadCluster
-	instance := &enterpriseApi.SearchHeadCluster{}
+	// Fetch the ClusterManager
+	instance := &enterpriseApi.ClusterManager{}
 	err := r.Get(ctx, req.NamespacedName, instance)
 	if err != nil {
 		if k8serrors.IsNotFound(err) {
@@ -89,20 +89,20 @@ func (r *SearchHeadClusterReconciler) Reconcile(ctx context.Context, req ctrl.Re
 			return ctrl.Result{}, nil
 		}
 		// Error reading the object - requeue the request.
-		return ctrl.Result{}, errors.Wrap(err, "could not load search head cluster data")
+		return ctrl.Result{}, errors.Wrap(err, "could not load cluster manager data")
 	}
 
 	// If the reconciliation is paused, requeue
 	annotations := instance.GetAnnotations()
 	if annotations != nil {
-		if _, ok := annotations[enterpriseApi.SearchHeadClusterPausedAnnotation]; ok {
+		if _, ok := annotations[enterpriseApi.ClusterManagerPausedAnnotation]; ok {
 			return ctrl.Result{Requeue: true, RequeueAfter: pauseRetryDelay}, nil
 		}
 	}
 
 	reqLogger.Info("start", "CR version", instance.GetResourceVersion())
 
-	result, err := ApplySearchHeadCluster(ctx, r.Client, instance)
+	result, err := ApplyClusterManager(ctx, r.Client, instance)
 	if result.Requeue && result.RequeueAfter != 0 {
 		reqLogger.Info("Requeued", "period(seconds)", int(result.RequeueAfter/time.Second))
 	}
@@ -110,46 +110,56 @@ func (r *SearchHeadClusterReconciler) Reconcile(ctx context.Context, req ctrl.Re
 	return result, err
 }
 
-// ApplySearchHeadCluster adding to handle unit test case
-var ApplySearchHeadCluster = func(ctx context.Context, client client.Client, instance *enterpriseApi.SearchHeadCluster) (reconcile.Result, error) {
-	return enterprise.ApplySearchHeadCluster(ctx, client, instance)
+// ApplyClusterManager adding to handle unit test case
+var ApplyClusterManager = func(ctx context.Context, client client.Client, instance *enterpriseApi.ClusterManager) (reconcile.Result, error) {
+	return enterprise.ApplyClusterManager(ctx, client, instance)
 }
 
 // SetupWithManager sets up the controller with the Manager.
-func (r *SearchHeadClusterReconciler) SetupWithManager(mgr ctrl.Manager) error {
+func (r *ClusterManagerReconciler) SetupWithManager(mgr ctrl.Manager) error {
 	return ctrl.NewControllerManagedBy(mgr).
-		For(&enterpriseApi.SearchHeadCluster{}).
+		For(&enterpriseApi.ClusterManager{}).
 		WithEventFilter(predicate.Or(
 			predicate.GenerationChangedPredicate{},
 			predicate.AnnotationChangedPredicate{},
 			common.LabelChangedPredicate(),
 			common.SecretChangedPredicate(),
-			common.ConfigMapChangedPredicate(),
 			common.StatefulsetChangedPredicate(),
 			common.PodChangedPredicate(),
+			common.ConfigMapChangedPredicate(),
+			common.CrdChangedPredicate(),
 		)).
 		Watches(&source.Kind{Type: &appsv1.StatefulSet{}},
 			&handler.EnqueueRequestForOwner{
 				IsController: false,
-				OwnerType:    &enterpriseApi.SearchHeadCluster{},
+				OwnerType:    &enterpriseApi.ClusterManager{},
 			}).
 		Watches(&source.Kind{Type: &corev1.Secret{}},
 			&handler.EnqueueRequestForOwner{
 				IsController: false,
-				OwnerType:    &enterpriseApi.SearchHeadCluster{},
-			}).
-		Watches(&source.Kind{Type: &corev1.ConfigMap{}},
-			&handler.EnqueueRequestForOwner{
-				IsController: false,
-				OwnerType:    &enterpriseApi.SearchHeadCluster{},
+				OwnerType:    &enterpriseApi.ClusterManager{},
 			}).
 		Watches(&source.Kind{Type: &corev1.Pod{}},
 			&handler.EnqueueRequestForOwner{
 				IsController: false,
-				OwnerType:    &enterpriseApi.SearchHeadCluster{},
+				OwnerType:    &enterpriseApi.ClusterManager{},
+			}).
+		Watches(&source.Kind{Type: &corev1.ConfigMap{}},
+			&handler.EnqueueRequestForOwner{
+				IsController: false,
+				OwnerType:    &enterpriseApi.ClusterManager{},
 			}).
 		WithOptions(controller.Options{
 			MaxConcurrentReconciles: enterpriseApi.TotalWorker,
 		}).
 		Complete(r)
+}
+
+// recordInstrumentionData Record api profiling information to prometheus
+func recordInstrumentionData(start time.Time, req ctrl.Request, module string, name string) {
+	metricLabels := getPrometheusLabels(req, name)
+	metricLabels[labelModuleName] = module
+	metricLabels[labelMethodName] = name
+	value := float64(time.Since(start) / time.Millisecond)
+	apiTotalTimeMetricEvents.With(metricLabels).Set(value)
 }
