+++ internal/controller/clustermanager_controller_test.go
@@ -1,14 +1,14 @@
-package controllers
+package controller
 
 import (
 	"context"
 	"fmt"
+	"github.com/splunk/splunk-operator/internal/controller/testutils"
 
 	enterpriseApi "github.com/splunk/splunk-operator/api/v4"
 
 	"time"
 
-	"github.com/splunk/splunk-operator/controllers/testutils"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
@@ -21,7 +21,7 @@ import (
 	"k8s.io/client-go/kubernetes/scheme"
 )
 
-var _ = Describe("MonitoringConsole Controller", func() {
+var _ = Describe("ClusterManager Controller", func() {
 
 	BeforeEach(func() {
 		time.Sleep(2 * time.Second)
@@ -31,56 +31,60 @@ var _ = Describe("MonitoringConsole Controller", func() {
 
 	})
 
-	Context("MonitoringConsole Management", func() {
+	Context("ClusterManager Management failed", func() {
 
-		It("Get MonitoringConsole custom resource should failed", func() {
-			namespace := "ns-splunk-mc-1"
-			ApplyMonitoringConsole = func(ctx context.Context, client client.Client, instance *enterpriseApi.MonitoringConsole) (reconcile.Result, error) {
+		It("Get ClusterManager custom resource should failed", func() {
+			namespace := "ns-splunk-cm-1"
+			ApplyClusterManager = func(ctx context.Context, client client.Client, instance *enterpriseApi.ClusterManager) (reconcile.Result, error) {
 				return reconcile.Result{}, nil
 			}
 			nsSpecs := &corev1.Namespace{ObjectMeta: metav1.ObjectMeta{Name: namespace}}
 			Expect(k8sClient.Create(context.Background(), nsSpecs)).Should(Succeed())
 			// check when resource not found
-			_, err := GetMonitoringConsole("test", nsSpecs.Name)
-			Expect(err.Error()).Should(Equal("monitoringconsoles.enterprise.splunk.com \"test\" not found"))
+			_, err := GetClusterManager("test", nsSpecs.Name)
+			Expect(err.Error()).Should(Equal("clustermanagers.enterprise.splunk.com \"test\" not found"))
 			Expect(k8sClient.Delete(context.Background(), nsSpecs)).Should(Succeed())
 		})
+	})
 
-		It("Create MonitoringConsole custom resource with annotations should pause", func() {
-			namespace := "ns-splunk-mc-2"
-			ApplyMonitoringConsole = func(ctx context.Context, client client.Client, instance *enterpriseApi.MonitoringConsole) (reconcile.Result, error) {
+	Context("ClusterManager Management with annotations", func() {
+
+		It("Create ClusterManager custom resource with annotations should pause", func() {
+			namespace := "ns-splunk-cm-2"
+			ApplyClusterManager = func(ctx context.Context, client client.Client, instance *enterpriseApi.ClusterManager) (reconcile.Result, error) {
 				return reconcile.Result{}, nil
 			}
 			nsSpecs := &corev1.Namespace{ObjectMeta: metav1.ObjectMeta{Name: namespace}}
 			Expect(k8sClient.Create(context.Background(), nsSpecs)).Should(Succeed())
 			annotations := make(map[string]string)
-			annotations[enterpriseApi.MonitoringConsolePausedAnnotation] = ""
-			CreateMonitoringConsole("test", nsSpecs.Name, annotations, enterpriseApi.PhaseReady)
-			ssSpec, _ := GetMonitoringConsole("test", nsSpecs.Name)
+			annotations[enterpriseApi.ClusterManagerPausedAnnotation] = ""
+			CreateClusterManager("test", nsSpecs.Name, annotations, enterpriseApi.PhaseReady)
+			ssSpec, _ := GetClusterManager("test", nsSpecs.Name)
 			annotations = map[string]string{}
 			ssSpec.Annotations = annotations
 			ssSpec.Status.Phase = "Ready"
-			UpdateMonitoringConsole(ssSpec, enterpriseApi.PhaseReady)
-			DeleteMonitoringConsole("test", nsSpecs.Name)
+			UpdateClusterManager(ssSpec, enterpriseApi.PhaseReady)
+			DeleteClusterManager("test", nsSpecs.Name)
 			Expect(k8sClient.Delete(context.Background(), nsSpecs)).Should(Succeed())
 		})
-
-		It("Create MonitoringConsole custom resource should succeeded", func() {
-			namespace := "ns-splunk-mc-3"
-			ApplyMonitoringConsole = func(ctx context.Context, client client.Client, instance *enterpriseApi.MonitoringConsole) (reconcile.Result, error) {
+	})
+	Context("ClusterManager Management", func() {
+		It("Create ClusterManager custom resource should succeeded", func() {
+			namespace := "ns-splunk-cm-3"
+			ApplyClusterManager = func(ctx context.Context, client client.Client, instance *enterpriseApi.ClusterManager) (reconcile.Result, error) {
 				return reconcile.Result{}, nil
 			}
 			nsSpecs := &corev1.Namespace{ObjectMeta: metav1.ObjectMeta{Name: namespace}}
 			Expect(k8sClient.Create(context.Background(), nsSpecs)).Should(Succeed())
 			annotations := make(map[string]string)
-			CreateMonitoringConsole("test", nsSpecs.Name, annotations, enterpriseApi.PhaseReady)
-			DeleteMonitoringConsole("test", nsSpecs.Name)
+			CreateClusterManager("test", nsSpecs.Name, annotations, enterpriseApi.PhaseReady)
+			DeleteClusterManager("test", nsSpecs.Name)
 			Expect(k8sClient.Delete(context.Background(), nsSpecs)).Should(Succeed())
 		})
 
 		It("Cover Unused methods", func() {
-			namespace := "ns-splunk-mc-4"
-			ApplyMonitoringConsole = func(ctx context.Context, client client.Client, instance *enterpriseApi.MonitoringConsole) (reconcile.Result, error) {
+			namespace := "ns-splunk-cm-4"
+			ApplyClusterManager = func(ctx context.Context, client client.Client, instance *enterpriseApi.ClusterManager) (reconcile.Result, error) {
 				return reconcile.Result{}, nil
 			}
 			nsSpecs := &corev1.Namespace{ObjectMeta: metav1.ObjectMeta{Name: namespace}}
@@ -88,7 +92,7 @@ var _ = Describe("MonitoringConsole Controller", func() {
 			ctx := context.TODO()
 			builder := fake.NewClientBuilder()
 			c := builder.Build()
-			instance := MonitoringConsoleReconciler{
+			instance := ClusterManagerReconciler{
 				Client: c,
 				Scheme: scheme.Scheme,
 			}
@@ -102,20 +106,18 @@ var _ = Describe("MonitoringConsole Controller", func() {
 			_, err := instance.Reconcile(ctx, request)
 			Expect(err).ToNot(HaveOccurred())
 			// create resource first and then reconcile for the first time
-			ssSpec := testutils.NewMonitoringConsole("test", namespace, "image")
+			ssSpec := testutils.NewClusterManager("test", namespace, "image")
 			Expect(c.Create(ctx, ssSpec)).Should(Succeed())
 			// reconcile with updated annotations for pause
 			annotations := make(map[string]string)
-			annotations[enterpriseApi.MonitoringConsolePausedAnnotation] = ""
+			annotations[enterpriseApi.ClusterManagerPausedAnnotation] = ""
 			ssSpec.Annotations = annotations
 			Expect(c.Update(ctx, ssSpec)).Should(Succeed())
 			_, err = instance.Reconcile(ctx, request)
-			Expect(err).ToNot(HaveOccurred())
 			// reconcile after removing annotations for pause
 			annotations = map[string]string{}
 			ssSpec.Annotations = annotations
 			Expect(c.Update(ctx, ssSpec)).Should(Succeed())
-			_, err = instance.Reconcile(ctx, request)
 			// reconcile after adding delete timestamp
 			Expect(err).ToNot(HaveOccurred())
 			ssSpec.DeletionTimestamp = &metav1.Time{}
@@ -126,13 +128,13 @@ var _ = Describe("MonitoringConsole Controller", func() {
 	})
 })
 
-func GetMonitoringConsole(name string, namespace string) (*enterpriseApi.MonitoringConsole, error) {
+func GetClusterManager(name string, namespace string) (*enterpriseApi.ClusterManager, error) {
 	key := types.NamespacedName{
 		Name:      name,
 		Namespace: namespace,
 	}
-	By("Expecting MonitoringConsole custom resource to be created successfully")
-	ss := &enterpriseApi.MonitoringConsole{}
+	By("Expecting ClusterManager custom resource to be created successfully")
+	ss := &enterpriseApi.ClusterManager{}
 	err := k8sClient.Get(context.Background(), key, ss)
 	if err != nil {
 		return nil, err
@@ -140,17 +142,17 @@ func GetMonitoringConsole(name string, namespace string) (*enterpriseApi.Monitor
 	return ss, err
 }
 
-func CreateMonitoringConsole(name string, namespace string, annotations map[string]string, status enterpriseApi.Phase) *enterpriseApi.MonitoringConsole {
+func CreateClusterManager(name string, namespace string, annotations map[string]string, status enterpriseApi.Phase) *enterpriseApi.ClusterManager {
 	key := types.NamespacedName{
 		Name:      name,
 		Namespace: namespace,
 	}
-	ssSpec := testutils.NewMonitoringConsole(name, namespace, "image")
+	ssSpec := testutils.NewClusterManager(name, namespace, "image")
 	Expect(k8sClient.Create(context.Background(), ssSpec)).Should(Succeed())
 	time.Sleep(2 * time.Second)
 
-	By("Expecting MonitoringConsole custom resource to be created successfully")
-	ss := &enterpriseApi.MonitoringConsole{}
+	By("Expecting ClusterManager custom resource to be created successfully")
+	ss := &enterpriseApi.ClusterManager{}
 	Eventually(func() bool {
 		_ = k8sClient.Get(context.Background(), key, ss)
 		if status != "" {
@@ -160,24 +162,24 @@ func CreateMonitoringConsole(name string, namespace string, annotations map[stri
 			time.Sleep(2 * time.Second)
 		}
 		return true
-	}, timeout, interval).Should(BeTrue())
+	}, NodeTimeout(timeout), interval).Should(BeTrue())
 
 	return ss
 }
 
-func UpdateMonitoringConsole(instance *enterpriseApi.MonitoringConsole, status enterpriseApi.Phase) *enterpriseApi.MonitoringConsole {
+func UpdateClusterManager(instance *enterpriseApi.ClusterManager, status enterpriseApi.Phase) *enterpriseApi.ClusterManager {
 	key := types.NamespacedName{
 		Name:      instance.Name,
 		Namespace: instance.Namespace,
 	}
 
-	ssSpec := testutils.NewMonitoringConsole(instance.Name, instance.Namespace, "image")
+	ssSpec := testutils.NewClusterManager(instance.Name, instance.Namespace, "image")
 	ssSpec.ResourceVersion = instance.ResourceVersion
 	Expect(k8sClient.Update(context.Background(), ssSpec)).Should(Succeed())
 	time.Sleep(2 * time.Second)
 
-	By("Expecting MonitoringConsole custom resource to be created successfully")
-	ss := &enterpriseApi.MonitoringConsole{}
+	By("Expecting ClusterManager custom resource to be created successfully")
+	ss := &enterpriseApi.ClusterManager{}
 	Eventually(func() bool {
 		_ = k8sClient.Get(context.Background(), key, ss)
 		if status != "" {
@@ -192,15 +194,15 @@ func UpdateMonitoringConsole(instance *enterpriseApi.MonitoringConsole, status e
 	return ss
 }
 
-func DeleteMonitoringConsole(name string, namespace string) {
+func DeleteClusterManager(name string, namespace string) {
 	key := types.NamespacedName{
 		Name:      name,
 		Namespace: namespace,
 	}
 
-	By("Expecting MonitoringConsole Deleted successfully")
+	By("Expecting ClusterManager Deleted successfully")
 	Eventually(func() error {
-		ssys := &enterpriseApi.MonitoringConsole{}
+		ssys := &enterpriseApi.ClusterManager{}
 		_ = k8sClient.Get(context.Background(), key, ssys)
 		err := k8sClient.Delete(context.Background(), ssys)
 		return err
