+++ internal/controller/clustermaster_controller_test.go
@@ -1,14 +1,15 @@
-package controllers
+package controller
 
 import (
 	"context"
 	"fmt"
+	"github.com/splunk/splunk-operator/internal/controller/testutils"
 
 	enterpriseApi "github.com/splunk/splunk-operator/api/v4"
 
 	"time"
 
-	"github.com/splunk/splunk-operator/controllers/testutils"
+	enterpriseApiV3 "github.com/splunk/splunk-operator/api/v3"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
@@ -21,7 +22,7 @@ import (
 	"k8s.io/client-go/kubernetes/scheme"
 )
 
-var _ = Describe("IndexerCluster Controller", func() {
+var _ = Describe("ClusterMaster Controller", func() {
 
 	BeforeEach(func() {
 		time.Sleep(2 * time.Second)
@@ -31,57 +32,60 @@ var _ = Describe("IndexerCluster Controller", func() {
 
 	})
 
-	Context("IndexerCluster Management", func() {
+	Context("ClusterMaster Management failed", func() {
 
-		It("Get IndexerCluster custom resource should failed", func() {
-			namespace := "ns-splunk-ic-1"
-			ApplyIndexerCluster = func(ctx context.Context, client client.Client, instance *enterpriseApi.IndexerCluster) (reconcile.Result, error) {
+		It("Get ClusterMaster custom resource should fail", func() {
+			namespace := "ns-splunk-cmaster-1"
+			ApplyClusterMaster = func(ctx context.Context, client client.Client, instance *enterpriseApiV3.ClusterMaster) (reconcile.Result, error) {
 				return reconcile.Result{}, nil
 			}
 			nsSpecs := &corev1.Namespace{ObjectMeta: metav1.ObjectMeta{Name: namespace}}
 			Expect(k8sClient.Create(context.Background(), nsSpecs)).Should(Succeed())
 			// check when resource not found
-			_, err := GetIndexerCluster("test", nsSpecs.Name)
-			Expect(err.Error()).Should(Equal("indexerclusters.enterprise.splunk.com \"test\" not found"))
+			_, err := GetClusterMaster("test", nsSpecs.Name)
+			Expect(err.Error()).Should(Equal("clustermasters.enterprise.splunk.com \"test\" not found"))
 			Expect(k8sClient.Delete(context.Background(), nsSpecs)).Should(Succeed())
 		})
+	})
 
-		It("Create IndexerCluster custom resource with annotations should pause", func() {
-			namespace := "ns-splunk-ic-2"
-			ApplyIndexerCluster = func(ctx context.Context, client client.Client, instance *enterpriseApi.IndexerCluster) (reconcile.Result, error) {
+	Context("ClusterMaster Management with annotations", func() {
+
+		It("Create ClusterMaster custom resource with annotations should pause", func() {
+			namespace := "ns-splunk-cmaster-2"
+			ApplyClusterMaster = func(ctx context.Context, client client.Client, instance *enterpriseApiV3.ClusterMaster) (reconcile.Result, error) {
 				return reconcile.Result{}, nil
 			}
 			nsSpecs := &corev1.Namespace{ObjectMeta: metav1.ObjectMeta{Name: namespace}}
 			Expect(k8sClient.Create(context.Background(), nsSpecs)).Should(Succeed())
 			annotations := make(map[string]string)
-			annotations[enterpriseApi.IndexerClusterPausedAnnotation] = ""
-			CreateIndexerCluster("test", nsSpecs.Name, annotations, enterpriseApi.PhaseReady)
-			ssSpec, _ := GetIndexerCluster("test", nsSpecs.Name)
+			annotations[enterpriseApiV3.ClusterMasterPausedAnnotation] = ""
+			CreateClusterMaster("test", nsSpecs.Name, annotations, enterpriseApi.PhaseReady)
+			ssSpec, _ := GetClusterMaster("test", nsSpecs.Name)
 			annotations = map[string]string{}
 			ssSpec.Annotations = annotations
 			ssSpec.Status.Phase = "Ready"
-			ssSpec.Status.ClusterManagerPhase, ssSpec.Status.ClusterMasterPhase = "Ready", "Ready" //CM* Phase can't be empty
-			UpdateIndexerCluster(ssSpec, enterpriseApi.PhaseReady)
-			DeleteIndexerCluster("test", nsSpecs.Name)
+			UpdateClusterMaster(ssSpec, enterpriseApi.PhaseReady)
+			DeleteClusterMaster("test", nsSpecs.Name)
 			Expect(k8sClient.Delete(context.Background(), nsSpecs)).Should(Succeed())
 		})
-
-		It("Create IndexerCluster custom resource should succeeded", func() {
-			namespace := "ns-splunk-ic-3"
-			ApplyIndexerCluster = func(ctx context.Context, client client.Client, instance *enterpriseApi.IndexerCluster) (reconcile.Result, error) {
+	})
+	Context("ClusterMaster Management", func() {
+		It("Create ClusterMaster custom resource should succeeded", func() {
+			namespace := "ns-splunk-cmaster-3"
+			ApplyClusterMaster = func(ctx context.Context, client client.Client, instance *enterpriseApiV3.ClusterMaster) (reconcile.Result, error) {
 				return reconcile.Result{}, nil
 			}
 			nsSpecs := &corev1.Namespace{ObjectMeta: metav1.ObjectMeta{Name: namespace}}
 			Expect(k8sClient.Create(context.Background(), nsSpecs)).Should(Succeed())
 			annotations := make(map[string]string)
-			CreateIndexerCluster("test", nsSpecs.Name, annotations, enterpriseApi.PhaseReady)
-			DeleteIndexerCluster("test", nsSpecs.Name)
+			CreateClusterMaster("test", nsSpecs.Name, annotations, enterpriseApi.PhaseReady)
+			DeleteClusterMaster("test", nsSpecs.Name)
 			Expect(k8sClient.Delete(context.Background(), nsSpecs)).Should(Succeed())
 		})
 
 		It("Cover Unused methods", func() {
-			namespace := "ns-splunk-ic-4"
-			ApplyIndexerCluster = func(ctx context.Context, client client.Client, instance *enterpriseApi.IndexerCluster) (reconcile.Result, error) {
+			namespace := "ns-splunk-cmaster-4"
+			ApplyClusterMaster = func(ctx context.Context, client client.Client, instance *enterpriseApiV3.ClusterMaster) (reconcile.Result, error) {
 				return reconcile.Result{}, nil
 			}
 			nsSpecs := &corev1.Namespace{ObjectMeta: metav1.ObjectMeta{Name: namespace}}
@@ -89,7 +93,7 @@ var _ = Describe("IndexerCluster Controller", func() {
 			ctx := context.TODO()
 			builder := fake.NewClientBuilder()
 			c := builder.Build()
-			instance := IndexerClusterReconciler{
+			instance := ClusterMasterReconciler{
 				Client: c,
 				Scheme: scheme.Scheme,
 			}
@@ -103,20 +107,18 @@ var _ = Describe("IndexerCluster Controller", func() {
 			_, err := instance.Reconcile(ctx, request)
 			Expect(err).ToNot(HaveOccurred())
 			// create resource first and then reconcile for the first time
-			ssSpec := testutils.NewIndexerCluster("test", namespace, "image")
+			ssSpec := testutils.NewClusterMaster("test", namespace, "image")
 			Expect(c.Create(ctx, ssSpec)).Should(Succeed())
 			// reconcile with updated annotations for pause
 			annotations := make(map[string]string)
-			annotations[enterpriseApi.IndexerClusterPausedAnnotation] = ""
+			annotations[enterpriseApiV3.ClusterMasterPausedAnnotation] = ""
 			ssSpec.Annotations = annotations
 			Expect(c.Update(ctx, ssSpec)).Should(Succeed())
 			_, err = instance.Reconcile(ctx, request)
-			Expect(err).ToNot(HaveOccurred())
 			// reconcile after removing annotations for pause
 			annotations = map[string]string{}
 			ssSpec.Annotations = annotations
 			Expect(c.Update(ctx, ssSpec)).Should(Succeed())
-			_, err = instance.Reconcile(ctx, request)
 			// reconcile after adding delete timestamp
 			Expect(err).ToNot(HaveOccurred())
 			ssSpec.DeletionTimestamp = &metav1.Time{}
@@ -127,13 +129,13 @@ var _ = Describe("IndexerCluster Controller", func() {
 	})
 })
 
-func GetIndexerCluster(name string, namespace string) (*enterpriseApi.IndexerCluster, error) {
+func GetClusterMaster(name string, namespace string) (*enterpriseApiV3.ClusterMaster, error) {
 	key := types.NamespacedName{
 		Name:      name,
 		Namespace: namespace,
 	}
-	By("Expecting IndexerCluster custom resource to be created successfully")
-	ss := &enterpriseApi.IndexerCluster{}
+	By("Expecting ClusterMaster custom resource to be created successfully")
+	ss := &enterpriseApiV3.ClusterMaster{}
 	err := k8sClient.Get(context.Background(), key, ss)
 	if err != nil {
 		return nil, err
@@ -141,23 +143,22 @@ func GetIndexerCluster(name string, namespace string) (*enterpriseApi.IndexerClu
 	return ss, err
 }
 
-func CreateIndexerCluster(name string, namespace string, annotations map[string]string, status enterpriseApi.Phase) *enterpriseApi.IndexerCluster {
+func CreateClusterMaster(name string, namespace string, annotations map[string]string, status enterpriseApi.Phase) *enterpriseApiV3.ClusterMaster {
 	key := types.NamespacedName{
 		Name:      name,
 		Namespace: namespace,
 	}
-	ssSpec := testutils.NewIndexerCluster(name, namespace, "image")
+	ssSpec := testutils.NewClusterMaster(name, namespace, "image")
 	Expect(k8sClient.Create(context.Background(), ssSpec)).Should(Succeed())
 	time.Sleep(2 * time.Second)
 
-	By("Expecting IndexerCluster custom resource to be created successfully")
-	ss := &enterpriseApi.IndexerCluster{}
+	By("Expecting ClusterMaster custom resource to be created successfully")
+	ss := &enterpriseApiV3.ClusterMaster{}
 	Eventually(func() bool {
 		_ = k8sClient.Get(context.Background(), key, ss)
 		if status != "" {
 			fmt.Printf("status is set to %v", status)
 			ss.Status.Phase = status
-			ss.Status.ClusterManagerPhase, ss.Status.ClusterMasterPhase = status, status //CM* Phase can't be empty
 			Expect(k8sClient.Status().Update(context.Background(), ss)).Should(Succeed())
 			time.Sleep(2 * time.Second)
 		}
@@ -167,25 +168,24 @@ func CreateIndexerCluster(name string, namespace string, annotations map[string]
 	return ss
 }
 
-func UpdateIndexerCluster(instance *enterpriseApi.IndexerCluster, status enterpriseApi.Phase) *enterpriseApi.IndexerCluster {
+func UpdateClusterMaster(instance *enterpriseApiV3.ClusterMaster, status enterpriseApi.Phase) *enterpriseApiV3.ClusterMaster {
 	key := types.NamespacedName{
 		Name:      instance.Name,
 		Namespace: instance.Namespace,
 	}
 
-	ssSpec := testutils.NewIndexerCluster(instance.Name, instance.Namespace, "image")
+	ssSpec := testutils.NewClusterMaster(instance.Name, instance.Namespace, "image")
 	ssSpec.ResourceVersion = instance.ResourceVersion
 	Expect(k8sClient.Update(context.Background(), ssSpec)).Should(Succeed())
 	time.Sleep(2 * time.Second)
 
-	By("Expecting IndexerCluster custom resource to be created successfully")
-	ss := &enterpriseApi.IndexerCluster{}
+	By("Expecting ClusterMaster custom resource to be created successfully")
+	ss := &enterpriseApiV3.ClusterMaster{}
 	Eventually(func() bool {
 		_ = k8sClient.Get(context.Background(), key, ss)
 		if status != "" {
 			fmt.Printf("status is set to %v", status)
 			ss.Status.Phase = status
-			ss.Status.ClusterManagerPhase, ss.Status.ClusterMasterPhase = status, status //CM* Phase can't be empty
 			Expect(k8sClient.Status().Update(context.Background(), ss)).Should(Succeed())
 			time.Sleep(2 * time.Second)
 		}
@@ -195,15 +195,15 @@ func UpdateIndexerCluster(instance *enterpriseApi.IndexerCluster, status enterpr
 	return ss
 }
 
-func DeleteIndexerCluster(name string, namespace string) {
+func DeleteClusterMaster(name string, namespace string) {
 	key := types.NamespacedName{
 		Name:      name,
 		Namespace: namespace,
 	}
 
-	By("Expecting IndexerCluster Deleted successfully")
+	By("Expecting ClusterMaster Deleted successfully")
 	Eventually(func() error {
-		ssys := &enterpriseApi.IndexerCluster{}
+		ssys := &enterpriseApiV3.ClusterMaster{}
 		_ = k8sClient.Get(context.Background(), key, ssys)
 		err := k8sClient.Delete(context.Background(), ssys)
 		return err
