+++ internal/controller/common/predicate.go
