+++ internal/controller/debug/register.go
