+++ internal/controller/metric.go
@@ -1,4 +1,4 @@
-package controllers
+package controller
 
 import (
 	"github.com/prometheus/client_golang/prometheus"
